# Remove loadToRAM method to improve startup performance